requirements.txt
amqp==5.0.6
asgiref==3.2.10
billiard==3.6.4.0
celery==5.1.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.1
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
rabbitmq==0.2.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
